Move lru-cache
